+++ src/components/UI/organisms/ApplicantCommentList.tsx
@@ -1,10 +1,12 @@
-import React from "react";
+import React, { useState } from "react";
 import styled from "styled-components";
 import { observer } from "mobx-react";
 
 import theme from "../../../styles/theme";
 import { IApplicantListData } from "../../../models/AdminAccountInterface";
+
 import AdminCommentForm from "../../../assets/data/adminCommentForm";
+import EvaluationBox from "../molecules/EvaluationBox";
 
 const CommentBox = styled.div`
   margin-top: 10px;
@@ -44,9 +46,19 @@ const AdminWrap = styled.div`
   flex-direction: column;
 `;
 
-const AdminAccount = styled.div`
-  margin-bottom: 5px;
+const TextWrap = styled.div`
   text-align: right;
+`;
+
+const EditComment = styled.span`
+  font-size: 10px;
+  color: ${theme.color.main};
+  cursor: pointer;
+`;
+
+const AdminAccount = styled.span`
+  margin-bottom: 5px;
+  margin-left: 10px;
   font-size: 6px;
 `;
 
@@ -64,35 +76,58 @@ const TextBox = styled.div`
 
 interface IAddData {
   item: IApplicantListData;
+  reRender?: any;
 }
 
-const ApplicantCommentList = observer(({ item }: IAddData): JSX.Element => {
-  const dateSlice = `${item.updated_at?.slice(0, 4)}/${item.updated_at?.slice(
-    5,
-    7
-  )}/${item.updated_at?.slice(8, 10)} `;
-  const meridiem = Number(item.updated_at?.slice(11, 13));
-  const timeSlice = ` ${item.updated_at?.slice(11, 16)}`;
-  const { score } = item;
+const ApplicantCommentList = observer(
+  ({ item, reRender }: IAddData): JSX.Element => {
+    const dateSlice = `${item.updated_at?.slice(0, 4)}/${item.updated_at?.slice(
+      5,
+      7
+    )}/${item.updated_at?.slice(8, 10)} `;
+    const meridiem = Number(item.updated_at?.slice(11, 13));
+    const timeSlice = ` ${item.updated_at?.slice(11, 16)}`;
 
-  return (
-    <CommentBox tabIndex={score}>
-      <CommentTitleWrap>
-        <TitleText tabIndex={score}>
-          {AdminCommentForm.evaluation[item.score].text}
-        </TitleText>
-        <AdminWrap>
-          <AdminAccount>{item.admin_name}</AdminAccount>
-          <TimeDate>
-            {dateSlice}
-            {meridiem >= 12 ? "오후" : "오전"}
-            {timeSlice}
-          </TimeDate>
-        </AdminWrap>
-      </CommentTitleWrap>
-      <TextBox>{item.description}</TextBox>
-    </CommentBox>
-  );
-});
+    const { score } = item;
+
+    const [isFormChange, setIsFormChange] = useState(true);
+    const handleEditComment = () => {
+      reRender();
+      setIsFormChange(!isFormChange);
+    };
+
+    return (
+      <>
+        <CommentBox tabIndex={score} hidden={!isFormChange}>
+          <CommentTitleWrap>
+            <TitleText tabIndex={score}>
+              {AdminCommentForm.evaluation[item.score].text}
+            </TitleText>
+            <AdminWrap>
+              <TextWrap>
+                <EditComment onClick={handleEditComment}>수정하기</EditComment>
+                <AdminAccount>{item.admin_name}</AdminAccount>
+              </TextWrap>
+              <TimeDate>
+                {dateSlice}
+                {meridiem >= 12 ? "오후" : "오전"}
+                {timeSlice}
+              </TimeDate>
+            </AdminWrap>
+          </CommentTitleWrap>
+          <TextBox>{item.description}</TextBox>
+        </CommentBox>
+        <EvaluationBox
+          score={score}
+          description={item.description}
+          checked={isFormChange}
+          edit={false}
+          commentId={item.id}
+          func={handleEditComment}
+        />
+      </>
+    );
+  }
+);
 
 export default ApplicantCommentList;
